Further route-loader abstractions
These changes add abstraction to the route-loading logic. Now, route
groups may be defined in the ./app/app.js file, under /routes, where the
first layer is the router-type and the second layer is the router file
for that type. The routes under /routes/http all register to the HTTP
Router (./lib/http/router.js).

In the future, these route groups will be used to differentiate how
different URIs should respond, in terms of middleware or restrictions.